Ceate a git repo struct
